+++ miniapp/src/test/java/com/rakuten/tech/mobile/miniapp/MiniAppDownloaderSpec.kt
@@ -326,6 +326,42 @@ class MiniAppDownloaderSpec : MiniAppDownloaderBaseSpec() {
 
             downloader.getMiniApp(TEST_ID_MINIAPP).first shouldBe TEST_BASE_PATH
         }
+    @Test
+    fun `getCachedMiniApp should return the miniapp path and miniapp info`() {
+        When calling miniAppStatus.isVersionDownloaded(
+            TEST_ID_MINIAPP,
+            TEST_ID_MINIAPP_VERSION,
+            TEST_BASE_PATH
+        ) itReturns true
+        When calling storage.getMiniAppVersionPath(
+            TEST_ID_MINIAPP,
+            TEST_ID_MINIAPP_VERSION
+        ) itReturns TEST_BASE_PATH
+        downloader.getCachedMiniApp(testMiniApp).first shouldBe TEST_BASE_PATH
+
+        When calling miniAppStatus.getDownloadedMiniApp(TEST_ID_MINIAPP) itReturns testMiniApp
+        downloader.getCachedMiniApp(TEST_ID_MINIAPP).first shouldBe TEST_BASE_PATH
+    }
+
+    @Test
+    fun `onGetCachedMiniApp should return the base path of the cached miniapp`() {
+        When calling storage.getMiniAppVersionPath(
+            TEST_ID_MINIAPP,
+            TEST_ID_MINIAPP_VERSION
+        ) itReturns TEST_BASE_PATH
+
+        downloader.onGetCachedMiniApp(testMiniApp).first shouldBe TEST_BASE_PATH
+    }
+
+    @Test(expected = MiniAppNotFoundException::class)
+    fun `should throw exception when cannot get miniapp by id from cache`() {
+        downloader.getCachedMiniApp(TEST_ID_MINIAPP)
+    }
+
+    @Test(expected = MiniAppNotFoundException::class)
+    fun `should throw exception when cannot get miniapp by miniapp info from cache`() {
+        downloader.getCachedMiniApp(testMiniApp)
+    }
 
     @Test(expected = MiniAppSdkException::class)
     fun `should throw exception when cannot get miniapp by id from server`() = runBlockingTest {