+++ testapp/src/main/java/com/rakuten/tech/mobile/testapp/ui/display/MiniAppDisplayViewModel.kt
@@ -41,10 +41,7 @@ class MiniAppDisplayViewModel constructor(
     ) = viewModelScope.launch(Dispatchers.IO) {
         try {
             _isLoading.postValue(true)
-            miniAppDisplay = if (appInfo != null)
-                miniapp.create(appInfo, miniAppMessageBridge, miniAppNavigator, miniAppFileChooser, appParameters)
-            else
-                miniapp.create(appId, miniAppMessageBridge, miniAppNavigator, miniAppFileChooser, appParameters)
+            miniAppDisplay = createMiniAppDisplay(appInfo, appId, miniAppMessageBridge, miniAppNavigator, miniAppFileChooser, appParameters)
             _miniAppView.postValue(miniAppDisplay.getMiniAppView(context))
         } catch (e: MiniAppSdkException) {
             e.printStackTrace()
@@ -53,12 +50,37 @@ class MiniAppDisplayViewModel constructor(
                     _errorData.postValue("No published version for the provided Mini App ID.")
                 is MiniAppNotFoundException ->
                     _errorData.postValue("No Mini App found for the provided Project ID.")
-                else -> _errorData.postValue(e.message)
+                else ->{
+                    //try to load MiniApp from cache
+                    try {
+                        miniAppDisplay = createMiniAppDisplay(appInfo, appId, miniAppMessageBridge, miniAppNavigator, miniAppFileChooser, appParameters, true)
+                        _miniAppView.postValue(miniAppDisplay.getMiniAppView(context))
+                    } catch (e: MiniAppSdkException) {
+                        when (e) {
+                            is MiniAppNotFoundException ->
+                                _errorData.postValue("No Mini App found for the provided Project ID.")
+                            else -> _errorData.postValue(e.message)
+                        }
+                    }
+                }
             }
         } finally {
             _isLoading.postValue(false)
         }
     }
+
+    private suspend fun createMiniAppDisplay(
+        appInfo: MiniAppInfo?,
+        appId: String,
+        miniAppMessageBridge: MiniAppMessageBridge,
+        miniAppNavigator: MiniAppNavigator,
+        miniAppFileChooser: MiniAppFileChooser,
+        appParameters: String,
+        fromCache: Boolean = false
+    ): MiniAppDisplay = if (appInfo != null)
+        miniapp.create(appInfo, miniAppMessageBridge, miniAppNavigator, miniAppFileChooser, appParameters, fromCache)
+    else
+        miniapp.create(appId, miniAppMessageBridge, miniAppNavigator, miniAppFileChooser, appParameters, fromCache)
 
     fun obtainMiniAppDisplayUrl(
         context: Context,